requirements.txt
jax
scipy
pytorch
numpy
matplotlib